refactor(resource): Rename assets/resources to assets/resourcedata
