+++ assets/resourcedata/resources.yaml
