+++ assets/resourcedata/slc-catalog.yaml
