+++ cmd/exasol/main.go
@@ -9,7 +9,7 @@ import (
 	"os"
 
 	// Activates embedded runtime resources for this binary.
-	_ "github.com/exasol/exasol-personal/assets/resources/generated"
+	_ "github.com/exasol/exasol-personal/assets/resourcedata/generated"
 	"github.com/exasol/exasol-personal/internal/util"
 )
