+++ internal/deploy/generated_resources_test.go
@@ -7,4 +7,4 @@ package deploy
 // package for its side effect (see cmd/exasol/main.go); tests need the same
 // import to resolve real named presets (e.g. "aws", "local") through the
 // resource cache instead of seeing an empty catalog.
-import _ "github.com/exasol/exasol-personal/assets/resources/generated"
+import _ "github.com/exasol/exasol-personal/assets/resourcedata/generated"
