+++ tools/resourceembedder/main_test.go
@@ -18,7 +18,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/exasol/exasol-personal/assets/resources"
+	"github.com/exasol/exasol-personal/assets/resourcedata"
 	"github.com/exasol/exasol-personal/internal/resource"
 )
 
@@ -684,7 +684,7 @@ func TestResolveResourceEmbed_RealPresetDirectoriesGlobToDeclaredMembers(t *test
 	if err != nil {
 		t.Fatalf("failed to resolve repository root: %v", err)
 	}
-	spec, err := resource.ParseSpec(resources.ResourcesYAML)
+	spec, err := resource.ParseSpec(resourcedata.ResourcesYAML)
 	if err != nil {
 		t.Fatalf("failed to parse resources.yaml: %v", err)
 	}
